refactor: reorganize UploadAbstraction structure and rename UploadInterface to UploaderInterface
