+++ composer.json
@@ -16,7 +16,8 @@
     },
     "autoload": {
         "psr-4": {
-            "UploadAbstractor\\": "src/"
+            "UploadAbstractor\\": "src/UploadAbstractor/",
+            "EnvManager\\": "src/EnvManager/"
         }
     },
     "require-dev": {
