+++ src/EnvManager/Readers/GetenvReader.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace UploadAbstractor\Support\Readers;
+namespace EnvManager\Readers;
 
-use UploadAbstractor\Support\Contracts\EnvironmentInterface;
+use EnvManager\Contracts\EnvironmentInterface;
 
 class GetenvReader implements EnvironmentInterface
 {
