+++ src/UploadAbstractor/Enums/UploaderDriver.php
@@ -1,7 +1,7 @@
 <?php
 namespace UploadAbstractor\Enums;
 
-enum UploadDriver: string
+enum UploaderDriver: string
 {
     case S3 = 's3';
     case LOCAL = 'local';
