+++ src/UploadAbstractor/Interfaces/UploaderInterface.php
@@ -1,7 +1,7 @@
 <?php
 namespace UploadAbstractor\Interfaces;
 
-interface UploadInterface
+interface UploaderInterface
 {
     public function createRepository(string $name): bool;
     public function upload(string $bucket, string $key, string $filePath): ?string;
